+++ R/resp.R
@@ -19,6 +19,8 @@
 #'   which will be parsed using the standard rules, or a named list.
 #' @param body Response, if any, contained in the response body.
 #'   For `response_json()`, a R data structure to serialize to JSON.
+#' @param timing A named numeric vector giving the time taken by various
+#'   components.
 #' @returns An HTTP response: an S3 list with class `httr2_response`.
 #' @export
 #' @examples
@@ -30,7 +32,8 @@ response <- function(
   url = "https://example.com",
   method = "GET",
   headers = list(),
-  body = raw()
+  body = raw(),
+  timing = NULL
 ) {
   check_number_whole(status_code, min = 100, max = 700)
   check_string(url)
@@ -43,7 +46,8 @@ response <- function(
     url = url,
     status_code = as.integer(status_code),
     headers = headers,
-    body = body
+    body = body,
+    timing = timing
   )
 }
 
@@ -76,6 +80,7 @@ new_response <- function(
   status_code,
   headers,
   body,
+  timing = NULL,
   request = NULL,
   error_call = caller_env()
 ) {
@@ -83,6 +88,9 @@ new_response <- function(
   check_string(url, call = error_call)
   check_number_whole(status_code, call = error_call)
   check_request(request, allow_null = TRUE)
+  if (!is.null(timing) && !is_bare_numeric(timing)) {
+    stop_input_type(timing, "a numeric vector", allow_null = TRUE)
+  }
 
   headers <- as_headers(headers, error_call = error_call)
   # ensure we always have a date field
@@ -97,6 +105,7 @@ new_response <- function(
       status_code = status_code,
       headers = headers,
       body = body,
+      timing = timing,
       request = request,
       cache = new_environment()
     ),
@@ -111,6 +120,7 @@ create_response <- function(req, curl_data, body) {
     status_code = curl_data$status_code,
     headers = as_headers(curl_data$headers),
     body = body,
+    timing = curl_data$times,
     request = req
   )