+++ NEWS.md
@@ -1,6 +1,7 @@
 # httr2 (development version)
 
 * `req_body_json_modify()` can now be used on a request with an empty body.
+* `resp_timing()` exposes timing information about the request measured by libcurl (@arcresu, #725).
 * `req_url_query()` now re-calculates n lengths when using `.multi = "explode"` to avoid select/recycling issues (@Kevanness, #719).
 
 # httr2 1.1.2